Struct jsonway::list_builder::ListBuilder

pub struct ListBuilder {
    pub null: bool,
    pub skip: bool,
    pub root: Option<String>,
    // some fields omitted
}

Fields

null

skip

root

Methods

impl ListBuilder

Use ListBuilder to produce JSON arrays

fn new() -> ListBuilder

fn from_json(list: Json) -> Option<ListBuilder>

Initialize builder with initial value.

fn build(builder: |&mut ListBuilder|) -> ListBuilder

Create new ListBuilder, pass it to closure as mutable ref and return.

fn push_json(&mut self, value: Json)

Push JSON value to list.

fn list(&mut self, builder: |&mut ListBuilder|)

Create new list and push it.

fn object(&mut self, builder: |&mut ObjectBuilder|)

Create new object and push it

fn null(&mut self)

It you call null, this list will be converted to null when converting to raw JSON value.

fn skip(&mut self)

It you call skip, this list will be skipped.

fn root(&mut self, root: &str)

fn has_root(&mut self) -> bool

fn unwrap(self) -> Json

Move out internal JSON value.

impl<T: ToJson> ListBuilder

fn push(&mut self, value: T)

Push to list something that can be converted to JSON.

impl<A, T: Iterator<A>> ListBuilder

fn objects(&mut self, iter: &mut T, func: |A, &mut ObjectBuilder|)

Fill this list by objects builded from iterator.

fn lists(&mut self, iter: &mut T, func: |A, &mut ListBuilder|)

fn map(&mut self, iter: &mut T, func: |A| -> Json)

Fill this list by JSON values builded from iterator.