Struct jsonway::array_builder::ArrayBuilder

pub struct ArrayBuilder {
    pub array: JsonArray,
    pub null: bool,
    pub skip: bool,
    pub root: Option<String>,
}

Fields
array: JsonArray null: bool skip: bool root: Option<String>
Methods

impl ArrayBuilder

Use ArrayBuilder to produce JSON arrays

pub fn new() -> ArrayBuilder

pub fn from_json(array: Value) -> Option<ArrayBuilder>

Initialize builder with initial value.

pub fn build<F>(builder: F) -> ArrayBuilder where
F: FnOnce(&mut ArrayBuilder),

Create new ArrayBuilder, pass it to closure as mutable ref and return.

pub fn push_json(&mut self, value: Value)

Push JSON value to array.

pub fn array<F>(&mut self, builder: F) where
F: FnOnce(&mut ArrayBuilder),

Create new array and push it.

pub fn object<F>(&mut self, builder: F) where
F: FnOnce(&mut ObjectBuilder),

Create new object and push it

pub fn null(&mut self)

It you call null, this array will be converted to null when converting to raw JSON value.

pub fn skip(&mut self)

It you call skip, this array will be skipped.

pub fn root(&mut self, root: &str)

pub fn has_root(&mut self) -> bool

pub fn unwrap(self) -> Value

Move out internal JSON value.

impl ArrayBuilder

pub fn push<T: Serialize>(&mut self, value: T)

Push to array something that can be converted to JSON.

impl ArrayBuilder

pub fn objects<A, T: Iterator<Item = A>, F>(&mut self, iter: T, func: F) where
F: Fn(A, &mut ObjectBuilder),

Fill this array by objects builded from iterator.

pub fn arrays<A, T: Iterator<Item = A>, F>(&mut self, iter: T, func: F) where
F: Fn(A, &mut ArrayBuilder),

pub fn map<A, T: Iterator<Item = A>, F>(&mut self, iter: T, func: F) where
F: Fn(A) -> Value,

Fill this array by JSON values builded from iterator.